TG-Bridger-master/requirements.txt
python-dotenv
telethon